Fix permissions
This commit renames the file_exchange app to fileexchange. This is done
because the app we use to apply permissions splits at underscores. Thus
many things break. This requires some shenanigans with the database, so
it is advised to just recreate it. Naming the app file-exchange wasn't
possible, because python apparently does not like importing modules with
dashes in them.

Some permissions also got renamed or removed, so please look at the
groups.yml file for further info. IMPORTANT: For some reason the
permission name specified in the file is wrong. You have to take the
name in front of the colon and add the substring after the point. Like
this:
~~course.add_course~~ => courses.add_course